+++ auth-service/services/rights_management_service.py
@@ -30,7 +30,7 @@ class RightsManagement:
         self.redis = redis
         self.session = session
 
-    async def create_right(self, right: CreateRightModel) -> RightModel:
+    async def create(self, right: CreateRightModel) -> RightModel:
         """Создание права в системе"""
         stmt = select(Right).where(Right.name == right.name)
         try:
@@ -42,27 +42,33 @@ class RightsManagement:
             await self.session.refresh(right_)
             return RightModel(id=right_.id, name=right_.name, description=right_.description)
         else:
-            raise ResponseError(massage=f"Право с названием '{right.name}' уже существует")
+            raise ResponseError(f"Право с названием '{right.name}' уже существует")
 
-    async def delete_right(self, right: SearchRightModel) -> str:
+    async def delete(self, right: SearchRightModel) -> str:
         """Удаление права в системе"""
         if not right.model_dump(exclude_none=True):
-            raise ResponseError(massage=NOT_ENOUGH_INFO)
+            raise ResponseError(NOT_ENOUGH_INFO)
 
         stmt = select(Right).where(or_(Right.name == right.name, Right.id == right.id))
         try:
             right_ = (await self.session.scalars(stmt)).one()
         except NoResultFound:
-            raise ResponseError(massage=f"Право '{right.name or right.id}' не существует")
+            raise ResponseError(f"Право '{right.name or right.id}' не существует")
         else:
+            if right_.name == 'admin':
+                raise ResponseError("Нельзя удалять право админа!")
             await self.session.delete(right_)
             await self.session.commit()
+            await self.redis.delete_right(right_.id)
             return f"Право '{right.name or right.id}' удалено"
 
-    async def change_right(self, right_old: SearchRightModel, right_new: ChangeRightModel) -> RightModel:
+    async def update(self, right_old: SearchRightModel, right_new: ChangeRightModel) -> RightModel:
         """Изменение свойств права в системе"""
         if not right_old.model_dump(exclude_none=True) or not right_new.model_dump(exclude_none=True):
-            raise ResponseError(massage=NOT_ENOUGH_INFO)
+            raise ResponseError(NOT_ENOUGH_INFO)
+        admin_right = await self.get_admin_right()
+        if right_old.id == admin_right.id or right_old.name == admin_right.name:
+            raise ResponseError("Нельзя изменять право админа!")
 
         stmt = (
             update(Right)
@@ -73,14 +79,14 @@ class RightsManagement:
         try:
             right_ = (await self.session.scalars(stmt)).one()
         except NoResultFound:
-            raise ResponseError(massage=f"Право '{right_old.name or right_old.id}' не существует")
+            raise ResponseError(f"Право '{right_old.name or right_old.id}' не существует")
         except IntegrityError:
-            raise ResponseError(massage=f"Право с названием '{right_new.name}' уже существует")
+            raise ResponseError(f"Право с названием '{right_new.name}' уже существует")
         else:
             await self.session.commit()
             return RightModel(id=right_.id, name=right_.name, description=right_.description)
 
-    async def get_all_rights(self) -> RightsModel:
+    async def get_all(self) -> RightsModel:
         """Выгрузка всех прав"""
         return RightsModel(
             rights=[
@@ -94,19 +100,19 @@ class RightsManagement:
         try:
             right_ = (await self.session.scalars(select(Right).where(Right.name == "admin"))).one()
         except NoResultFound:
-            raise ResponseError(massage="Право 'admin' не существует")
+            raise ResponseError("Право 'admin' не существует")
         return right_
 
-    async def assign_user_right(self, right: SearchRightModel, user: UserModel) -> ResponseUserModel:
+    async def assign(self, right: SearchRightModel, user: UserModel) -> ResponseUserModel:
         """Назначение права пользователю"""
         if not right.model_dump(exclude_none=True) or not user.model_dump(exclude_none=True):
-            raise ResponseError(massage=NOT_ENOUGH_INFO)
+            raise ResponseError(NOT_ENOUGH_INFO)
 
         stmt_right = select(Right).where(or_(Right.name == right.name, Right.id == right.id))
         try:
             right_ = (await self.session.scalars(stmt_right)).one()
         except NoResultFound:
-            raise ResponseError(massage=f"Право '{right.name or right.id}' не существует")
+            raise ResponseError(f"Право '{right.name or right.id}' не существует")
 
         stmt_user = (
             select(User)
@@ -116,11 +122,11 @@ class RightsManagement:
         try:
             user_ = (await self.session.scalars(stmt_user)).one()
         except NoResultFound:
-            raise ResponseError(massage=f"Пользователь '{user.id or user.login or user.email}' не существует")
+            raise ResponseError(f"Пользователь '{user.id or user.login or user.email}' не существует")
 
         if right_ in user_.rights:
             raise ResponseError(
-                massage=(
+                message=(
                     f"Пользователь '{user.id or user.login or user.email}' "
                     f"уже имеет право '{right.name or right.id}'"
                 )
@@ -139,16 +145,16 @@ class RightsManagement:
         await self.redis.add_user_right(user_.id, right_.id)
         return result
 
-    async def take_away_right(self, right: SearchRightModel, user: UserModel) -> ResponseUserModel:
+    async def take_away(self, right: SearchRightModel, user: UserModel) -> ResponseUserModel:
         """Отъём права у пользователя"""
         if not right.model_dump(exclude_none=True) or not user.model_dump(exclude_none=True):
-            raise ResponseError(massage=NOT_ENOUGH_INFO)
+            raise ResponseError(NOT_ENOUGH_INFO)
 
         stmt_right = select(Right).where(or_(Right.name == right.name, Right.id == right.id))
         try:
             right_ = (await self.session.scalars(stmt_right)).one()
         except NoResultFound:
-            raise ResponseError(massage=f"Право '{right.name or right.id}' не существует")
+            raise ResponseError(f"Право '{right.name or right.id}' не существует")
 
         stmt_user = (
             select(User)
@@ -158,13 +164,13 @@ class RightsManagement:
         try:
             user_ = (await self.session.scalars(stmt_user)).one()
         except NoResultFound:
-            raise ResponseError(massage=f"Пользователь '{user.id or user.login or user.email}' не существует")
+            raise ResponseError(f"Пользователь '{user.id or user.login or user.email}' не существует")
 
         try:
             user_.rights.remove(right_)
         except ValueError:
             raise ResponseError(
-                massage=(
+                (
                     f"Пользователь '{user.id or user.login or user.email}' "
                     f"не имеет право '{right.name or right.id}'"
                 )
@@ -185,7 +191,7 @@ class RightsManagement:
     async def get_user_rights(self, user: UserModel) -> RightsModel:
         """Выгрузка всех прав пользователя"""
         if not user.model_dump(exclude_none=True):
-            raise ResponseError(massage=NOT_ENOUGH_INFO)
+            raise ResponseError(NOT_ENOUGH_INFO)
 
         stmt = (
             select(User)
@@ -195,7 +201,7 @@ class RightsManagement:
         try:
             rights = (await self.session.scalars(stmt)).unique().one()
         except NoResultFound:
-            raise ResponseError(massage=f"Пользователь '{user.id or user.login or user.email}' не существует")
+            raise ResponseError(f"Пользователь '{user.id or user.login or user.email}' не существует")
 
         return RightsModel(
             rights=[RightModel(id=right.id, name=right.name, description=right.description) for right in rights.rights]
